services/ml/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
joblib==1.4.2
shap==0.46.0

psycopg[binary]==3.2.12
python-ulid==2.7.0
python-dotenv==1.0.1
